docs(Root): delete helper, create utils, move archive
