+++ app/interface/page.jsx
@@ -1,7 +1,7 @@
 import ImageHero from "@/components/componentspages/imagehero";
 import Footer from "@/components/footer/footer";
 import Header from "@/components/header/header";
-import LoadingWrapper from "@/components/helper/loadingwrapper";
+import LoadingWrapper from "@/components/utils/loadingwrapper";
 import InterfaceDescription from "./components/interfacedescription";
 import GalleryVideo from "./components/galleryvideo";
