+++ components/utils/loader.module.scss
